moma_demos/articulated_demo/requirements.txt
pybullet~=3.0.1
scipy~=1.5.2
numpy~=1.19.2
PyYAML
matplotlib